Feat: 헤더 레이아웃 추가 작업사항 - 헤더 레이아웃 추가 - 헤더에 사용할 이미지 파일 추가 - 메인 액티비티 수정
