+++ .gitignore
@@ -63,7 +63,7 @@ proguard/
 
 # Android Studio
 /*/build/
-/*/local.properties
+/local.properties
 /*/out
 /*/*/build
 /*/*/production
