+++ app/src/main/java/com/diary/paintlog/view/activities/MainActivity.kt
@@ -21,7 +21,6 @@ import java.time.LocalDateTime
 
 class MainActivity : AppCompatActivity() {
 
-    private lateinit var appBarConfiguration: AppBarConfiguration
     private lateinit var binding: ActivityMainBinding
 
     companion object {
@@ -33,93 +32,5 @@ class MainActivity : AppCompatActivity() {
 
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
-
-        setSupportActionBar(binding.toolbar)
-
-        val dataStore = (application as GlobalApplication).dataStore
-
-        val navController = findNavController(R.id.nav_host_fragment_content_main)
-        appBarConfiguration = AppBarConfiguration(navController.graph)
-        setupActionBarWithNavController(navController, appBarConfiguration)
-
-        binding.fab.setOnClickListener { view ->
-            val repo = TokenRepository(dataStore)
-            CoroutineScope(Dispatchers.Main).launch {
-
-                Log.i(TAG, repo.getToken())
-                Log.i(TAG, repo.getTokenExpireAt())
-                Log.i(TAG, repo.getRefreshToken())
-                Log.i(TAG, repo.getRefreshTokenExpireAt())
-
-//                Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
-//                    .setAction("Action", null).show()
-            }
-        }
-
-//        Kakao.openKakaoLogin(binding.root.context)
-
-        val userDao = GlobalApplication.database.userDao()
-
-        var user1 = User(
-            1L,
-            "REGISTERED",
-            134513L,
-            135315L,
-            LocalDateTime.now(),
-            LocalDateTime.now(),
-            LocalDateTime.now()
-        )
-        var user2 = User(
-            2L,
-            "REGISTERED",
-            134513L,
-            135315L,
-            LocalDateTime.now(),
-            LocalDateTime.now(),
-            LocalDateTime.now()
-        )
-        var user3 = User(
-            3L,
-            "REGISTERED",
-            134513L,
-            135315L,
-            LocalDateTime.now(),
-            LocalDateTime.now(),
-            LocalDateTime.now()
-        )
-
-        // 코루틴을 사용한 예시
-        CoroutineScope(Dispatchers.IO).launch {
-//            userDao.insertAll(user1, user2, user3, user4, user5, user6, user7)
-
-            val users = userDao.getAll()
-
-            for (user in users) {
-                Log.i(TAG, user.toString())
-            }
-        }
-
-    }
-
-    override fun onCreateOptionsMenu(menu: Menu): Boolean {
-        // Inflate the menu; this adds items to the action bar if it is present.
-        menuInflater.inflate(R.menu.menu_main, menu)
-        return true
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        // Handle action bar item clicks here. The action bar will
-        // automatically handle clicks on the Home/Up button, so long
-        // as you specify a parent activity in AndroidManifest.xml.
-        return when (item.itemId) {
-            R.id.action_settings -> true
-            else -> super.onOptionsItemSelected(item)
-        }
-    }
-
-    override fun onSupportNavigateUp(): Boolean {
-        val navController = findNavController(R.id.nav_host_fragment_content_main)
-        return navController.navigateUp(appBarConfiguration)
-                || super.onSupportNavigateUp()
     }
 }
